+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## 0.2.0 - Feb 2 2026
 
+### Fixed
+
+- Fixed grammar to support optional `int` typing in function parameters.
+  - `function add(int a, int b)...`
+- Fixed grammar to support `identifier`s in array sizes and enum assignment.
+  - `enum { VAR = CONSTANT, ...}`, `int Arr[CONSTANT];`
+- Warn when a function parameter is being overshadowed by a declared variable.
+
 ### Added
 
 - Errors when trying to access a variable you havent defined yet.